Add function to automatically detect input files
The function f90-buffer-is-input-file determines if the current buffer is a
Fortran input file. It returns true if it is. Instructions for running this
function on opening a buffer are included in the readme file.

f90-buffer-is-input-file works by determining if the first non-blank, non-comment
line is a Fortran namelist, although it does this in a pretty basic fashion, by
just looking if the first word is "/" or "&[a-zA-Z1-9_]+". Some tweaking of
f90-namelist-startreg is probably required.